refactor: move repository definitions to domain
