+++ src/db/sqlx/payment_repository.rs
@@ -4,7 +4,7 @@ use sqlx::{query, QueryBuilder, Sqlite, SqlitePool};
 use crate::config::Config;
 use crate::db::sqlx::criteria::SqlxCriterion;
 use crate::db::sqlx::record::IncomeRecord;
-use crate::db::PaymentRepository;
+use crate::domain::repository::PaymentRepository;
 use crate::domain::Income;
 use crate::domain::Payment;
 use crate::income::criteria::{IncomeCriteria, IncomeCriterion};
