+++ src/domain/repository/payment.rs
@@ -1,0 +1,17 @@
+//! Database support.
+//! The database keeps incomes to query them later.
+//! It can be used as a data source instead of reading bank CSV every time.
+
+use async_trait::async_trait;
+
+use crate::domain::Payment;
+use crate::income::criteria::IncomeCriteria;
+
+#[async_trait]
+pub trait PaymentRepository {
+    async fn find_by(&mut self, criteria: IncomeCriteria) -> anyhow::Result<Vec<Payment>>;
+
+    async fn mark_paid(&mut self, payment_no: i64) -> anyhow::Result<()>;
+
+    async fn mark_unpaid(&mut self, payment_no: i64) -> anyhow::Result<()>;
+}
